requirements.txt
# Delta Exchange Trading Bot Requirements

# Core Dependencies
requests>=2.31.0
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Data Processing
pandas>=2.1.0
numpy>=1.24.0

# FastAPI and Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
python-multipart>=0.0.6

# Database (for trade history)
sqlalchemy>=2.0.23
aiosqlite>=0.19.0

# Logging and Monitoring
colorlog>=6.8.0

# Optional: Notifications
# python-telegram-bot>=20.7  # Uncomment if using Telegram notifications

# Development Tools (optional)
# pytest>=7.4.3
# pytest-asyncio>=0.21.1
# black>=23.11.0
# flake8>=6.1.0
